Move predicate test files into their own package
