+++ docs/DeveloperGuide.md
@@ -300,7 +300,7 @@ Priorities: High (must have) - `* * *`, Medium (nice to have) - `* *`, Low (unli
 | `* *`    | user                                                  | search for clients by name, phone number, or tag                         | quickly retrieve their information without scrolling through the entire list   |
 | `* *`    | salesperson                                           | add profile pictures for my clients                                      | quickly recognize them and personalize my interactions                         |
 | `* *`    | salesperson                                           | add social media links to client profiles                                | easily connect with them on different platforms                                |
-| `* *`    | salesperson whose clients are of different statuses   | categorize clients into different groups (e.g., VIP, Potential, Regular) | manage them more efficiently and tailor my communication                       |  
+| `* *`    | salesperson whose clients are of different statuses   | categorize clients into different groups (e.g., VIP, Potential, Regular) | manage them more efficiently and tailor my communication                       |
 | `* *`    | salesperson whose customers come from many industries | add tags for clients                                                     | split them into categories                                                     |
 | `* *`    | salesperson                                           | filter clients by tags                                                   | quickly see all the clients inside one category                                |
 | `* *`    | user                                                  | be able to use the software in dark mode                                 | view the contents comfortably                                                  |
