+++ src/test/java/seedu/address/model/client/predicates/CategoryContainsKeywordPredicateTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.client;
+package seedu.address.model.client.predicates;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -10,7 +10,6 @@ import java.util.Collections;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
-import seedu.address.model.client.predicates.CategoryContainsKeywordsPredicate;
 import seedu.address.testutil.ClientBuilder;
