+++ src/test/java/stub/model/client/NameContainsKeywordsPredicateStub.java
@@ -20,5 +20,4 @@ public class NameContainsKeywordsPredicateStub extends NameContainsKeywordsPredi
     public boolean test(Client client) {
         return returnValue;
     }
-
 }
